test: add initial set of unit tests and mock data
