--- test/fixtures/issues.opened.json
@@ -1,18 +1,0 @@
-{
-  "action": "opened",
-  "issue": {
-    "number": 1,
-    "user": {
-      "login": "hiimbex"
-    }
-  },
-  "repository": {
-    "name": "testing-things",
-    "owner": {
-      "login": "hiimbex"
-    }
-  },
-  "installation": {
-    "id": 2
-  }
-}
